Moved definition of gtypes and term data structures from module Gtypes and module Term to module Basic.
Added quantifier kind to Basic.Qnt, making explicit the type of the
quantification (lambda, exists or forall).
In case of conflict between quantifier kind in bound variable and
binding term Qnt, the binding term wins.
I.e. if in Qnt(k, q, _), (binder_kind q) is not k then the term is not
well formed.