Changes to subgoals.
Added type subgoal and made Goal a list of subgoals.
Each subgoal is a list of sequents.

This makes a goal a stack of tree levels, with each subgoal formed
from the application of a rule to a sequent.

(Previously, a goal was a list of sequents to be proved, with no way
of determining the subgoals resulting from the application of a rule.)