New subgoals package.
A tactic is now a node -> branch.
A node is a single sequent, a branch a list of sequents.
Module Logic.Subgoals has functions to apply tactics to
nodes and branch and merge the resulting subgoals.

These changes mean that complex tactics needed to be reworked. Since
the most complex tactics don't work anyway, this shouldn't be a problem.