Changed handling of skolem constants to avoid possible soundness bug.
A skolem constant is now a meta term, a bound variable with
quantifier Meta. A meta term is treated like a term identifier: meta
variables are added to the scope by the skolem functions. A term is
closed in a scope if all meta terms occuring in the term are in scope.
For closed term tests which do not take a scope, the meta term is
assumed to always be bound.

The changes uncovered some bugs including one in Net.insert in which
rules were dropped instead of added to the net.