[Makefiles] Don't use staging directory when building.
The staging directory was used to collect together the objects that were
generated as part of the build. This was needed because different parts
of the build needed access to objects generated as part of the
build. The staging directory was used as a single point which any part
of the build could access.

The staging area made the build system unnecessarily complicated. It is
removed and the build rearranged to access generated objects in the
directory in which they are originally created.