[configure, Makefiles] Rearrange installation directories.
The theories files were installed in the libraries directory in the
directory <prefix>/lib/thys. This wasn't appropriate so they are now
installed into a data directory for hseq, <prefix>/share/hseq/thys.

The documentation files are also moved so that they are now installed
into <prefix>/share/hseq/doc.

A new option --datadir is added to the configure script to allow the
data directory to be explicitly set.