[hsequser] Properly load theory library files.
Theory library files are intended to be able to update the system with by
loading and executing code. This is used to provide the printers and parsers
for the Pair and Set theories. For this to work, the library files need to be
able to update the global HSeq state. Because the theory loader did not do
support this, the effect of loading library files was lost.

This is fixed by having the theory loader updating the global state when
loading library files.

When library files specified by a theory were loaded, their effects were not
properly registered. This was due to theory libraries being the theory loader not updating the global state