Generate and use build configuration data as ocaml sources.
The build configure script generated configuration data as a set of macro
definitions suitable for use with a camlp4 based preprocessor. This made the
build dependent on the preprocessor.

The dependence on the preprocessor has been removed. The build configure
script now generates config.{mli,ml} files suitable for compiling and linking
into the libraries. The hseq and tools/scripts build are updated to use these
files.