[hseq] Remove dependency on build-time quotation system.
The build system relies on campl4 to provide a quotation system for embedding
HSeq terms and types in the Ocaml code. This can be difficult to maintain and
it is better to reduce the dependency.

The hseq source directory files are now no longer dependent on the quotation
system.

Also some whitespace cleanup and updated copyright headers.