[thys] Fix build failure.
Building the theories involves creating a multi-line OCaml file
'local_thy_config.ml' with the tempory paths configuration. This was done
using the echo shell command to write a single string to the file, with
control characters to output the newlines. This doesn't work on systems where
echo is not set up to interpret control characters by default, causing the
build to failure since the resulting file is not valid OCaml.

The configuration file for the theories build is now created with one
invocation of 'echo' per output line.