Rework handling of configuration settings.
The configure script used to write an OCaml file, config.{mli,ml}, with the
configuration settings. This was copied into various directories as part of
the build which lead to multiple copies of the same information.

The configure script now writes the OCaml files into the hseq directory where
they are built into the hseq.{cma,cmxa} library code as module Config and code
depending on the configuration values now links in the hseq libraries.